refactor: make it actually work
The plugins providing any rules which any of the configs utilize (e.g. react-hooks) need to be
included even if they aren't relevant to every config, else errors are thrown when linting because
the rules can't be found. It's not perfect, because all the rules called out in the various configs
are being run regardless of the config being used, but at least it's only those and not the entire
ruleset provided by each plugin (e.g. react/jsx-filename-extension is run, but not all the other
react/* rules). Ideally, this should be improved/fixed, but it's tolerable for now. Also, there is a
chance the resolution of bundled plugins is only working due to Yarn (flatter node_modules) and
ESLint hoisting (which is unreliable and risky). But this should suffice until ESLint has a proper
solution for including plugins as dependencies of shareable configs (work ongoing).