feat(rules): disable @typescript-eslint/explicit-function-return-type
The most important thing is enforcing explicit function return types on exported functions.
That scenario is covered by @typescript-eslint/explicit-module-boundary-types, so this rule
can be disabled to allow TypeScript to infer return types of internal functions (which can
still be declared explicitly if needed), which should provide a smoother developer experience.